Error-based turning
Decides whether to turn or spin/vector based on position, then decides
spin or vector based on error. This ends up being a bad way to do it...
basing the vector/spin decision on position works much more reliably and
gets the robot much straighter on the line following a spin.